Add Cache2Db Class with MD5 Hash function, set Cache2Db Project, and CSVDataMasking Unit test
In the Cache2Db class, added a function to compute MD5 hash of the given string payload and clarify its cryptographic limitations. New class Cache2Db has a loadWordList function to add to a Dictionary, and a function to lookup a word in the Dictionary. CSVDataMasking Unit tests were added to test these functionalities. The intention of the changes is to use these mechanisms to mask specific data fields in a CSV file. Updated the CSVDataMasking.sln solution file to include the new Cache2Db project and the CSVDataMasking Unit Test project.