Add serialization to cache
In this commit, we have added functionality to save and read the cache (the 'hitlist') to and from a file, enabling persistency of masked words between runs. It saves the state as a JSON file when the CSV writer finishes its work, and reads it when loading the word list. This is done via the new methods 'SaveDictionaryToFile' and 'ReadDictionaryFromFile' in the Cache class. This change will reduce unnecessary computations and increase the efficiency of the system.